Modification déplacement ship1 en avant arrière tourner gauche tourner droite

+++ src/main/java/fr/cpe/model/AllyShip.java
@@ -5,7 +5,7 @@ public class AllyShip extends SpacialObject {
     public int maxHp;
     public int reloadTime;
 
-    public AllyShip(double x, double y, int direction, int speed) {
+    public AllyShip(double x, double y, double direction, int speed) {
         super(x, y, direction, speed);
     }
 }
