+++ src/main/java/fr/cpe/model/Enemy.java
@@ -8,7 +8,7 @@ abstract public class Enemy extends SpacialObject implements EnemyStrategy {
     public int maxHp;
     public int reloadTime;
 
-    public Enemy(double x, double y, int direction, int speed) {
+    public Enemy(double x, double y, double direction, int speed) {
         super(x, y, direction, speed);
     }
