+++ src/main/java/fr/cpe/model/Projectile.java
@@ -1,7 +1,7 @@
 package fr.cpe.model;
 
 abstract public class Projectile extends SpacialObject{
-    public Projectile(double x, double y, int direction, int speed) {
+    public Projectile(double x, double y, double direction, int speed) {
         super(x, y, direction, speed);
     }
 }
